implemented entity anchor argument type
