+++ pumpkin/src/command/argument_types/mod.rs
@@ -189,6 +189,7 @@ pub mod argument_type;
 pub mod coordinates;
 pub mod core;
 pub mod entity;
+pub mod entity_anchor;
 pub mod entity_selector;
 pub mod game_profile;
 pub mod range;
